refactor(hooks): extract chat-stream sub-hooks into focused files
Per rubber-duck-reviewed Phase 6 plan, promote the three already-isolated
inline functions in use-learning-chat-stream.ts (616 LOC) into named
files. No behaviour change — the file already contained the right
seams; this exposes them as importable modules.

- chat-stream-events.ts: SSEStreamEvent type, buildTerminalHandler,
  dispatchFrame. Pure dispatch helpers, no React.
- use-chat-sse-subscriptions.ts: opens one SSE per active chat job,
  surfaces reconnecting state. Chat-specific (rubber-duck flagged that
  consumeSSE is already the generic primitive, so no useSSESubscription).
- use-pending-stream-reconciler.ts: pendingStreamMessages reconciler.
- use-orphaned-chat-stream-finalizer.ts: STALE_STREAM_THRESHOLD safety
  net, renamed from useStaleStreamFinalizer per rubber-duck (the
  prior name implied heartbeat detection; behaviour is orphan-thread
  cleanup, not stale-stream reconnect).

use-learning-chat-stream.ts is now a 284-LOC coordinator that owns
the cross-cutting derivations (streamingThreadIdsKey,
chatSubscriptionKey, isStreaming) and the public action surface. All
38 chat-stream tests pass unchanged.

Baseline: use-learning-chat-stream removed from production budget.
